Create Table 4.3

Live births by place of occurrence and place of usual residence of mother, year

Variables required:

Date of occurrence (birth1a)

Place of occurrence (birth1c)

Place of usual residence - Mother (birth3l)

Usage

create_t4.3(data, date_var, data_year = 2022, tablename = "Table_4_3")

Arguments

data

Birth registration data

date_var

Date of occurrence variable

data_year

Chosen year

tablename

Name for csv output use _ instead of . for names

Value

Data frame with tabulated results

Note

The Vital Statistics report resource kit includes guidance and templates to help complete your CRVS report.

Examples

table4_3 <- create_t4.3(sample_data1, date_var = birth1a, data_year = 2021, tablename = "Table_4_3")
#> Error in na.omit(data, sbind): object 'sample_data1' not found